HSEARCH-746 Move InfinispanDirectoryProvider to impl
