+++ hibernate-search-infinispan/src/test/java/org/hibernate/search/infinispan/ClusterTestHelper.java
@@ -25,6 +25,7 @@ import junit.framework.AssertionFailedError;
 
 import org.hibernate.cfg.Environment;
 import org.hibernate.search.SearchFactory;
+import org.hibernate.search.infinispan.impl.InfinispanDirectoryProvider;
 import org.hibernate.search.store.DirectoryProvider;
 import org.hibernate.search.test.util.FullTextSessionBuilder;
 import org.infinispan.manager.EmbeddedCacheManager;
@@ -96,7 +97,7 @@ public class ClusterTestHelper {
 	public static int clusterSize(FullTextSessionBuilder node) {
 		SearchFactory searchFactory = node.getSearchFactory();
 		DirectoryProvider[] directoryProviders = searchFactory.getDirectoryProviders( SimpleEmail.class );
-		InfinispanDirectoryProvider directoryProvider = (InfinispanDirectoryProvider) directoryProviders[0];
+		InfinispanDirectoryProvider directoryProvider = (InfinispanDirectoryProvider ) directoryProviders[0];
 		EmbeddedCacheManager cacheManager = directoryProvider.getCacheManager();
 		List<Address> members = cacheManager.getMembers();
 		return members.size();
